Decluttering generate.py
