+++ graphics/app/generate.py
@@ -6,69 +6,17 @@ from PIL import Image, ImageDraw, ImageFont
 # This is only used for type hinting
 from PIL.Image import Image as ImageType
 
+from graphics.utils.CharacterImage import (get_character_path, generate_character_image)
+from graphics.utils.Text import biggest_font_size
 from graphics.app.types import TournamentData
-from graphics.definitions import FONTS_DIR, OUTPUT_DIR, RENDERS_DIR, THUMBNAIL_DIR
+from graphics.definitions import FONTS_DIR, OUTPUT_DIR, THUMBNAIL_DIR
 
 SIZE = (1920, 1080)
 CHARACTER_BOX = (960, 800)
 POSITION = [(0, 140), (960, 140)]
 
 
-def get_character_path(character: str, alt: str = "01") -> str:
-    if character.lower() == "random":
-        return os.path.join(RENDERS_DIR, "Random/Random.png")
-    # The Mii Fighters only have one alt.
-    if "mii" in character.lower():
-        alt = "01"
-
-    return os.path.join(RENDERS_DIR, f"{character}/{alt}.png")
-
-
-def crop_character(
-    image: ImageType, crop: tuple[int, int], offset: tuple[int, int]
-) -> ImageType:
-    img_width, img_height = image.size
-
-    return image.crop(
-        (
-            (img_width - crop[0] + offset[0]) // 2,
-            (img_height - crop[1] + offset[1]) // 2,
-            (img_width + crop[0] + offset[0]) // 2,
-            (img_height + crop[1] + offset[1]) // 2,
-        )
-    )
-
-
-def resize_character(image: ImageType, zoom: int) -> ImageType:
-    width_one, height_one = image.size
-    if width_one > 960:
-        factor = 960 / width_one
-        width_one = int(width_one * factor * zoom / 100)
-        height_one = int(height_one * factor * zoom / 100)
-    return image.resize((width_one, height_one))
-
-
-def generate_character_image(
-    path: str, offset: tuple[int, int], zoom: int
-) -> ImageType:
-    character = Image.open(path, mode="r")
-    character = character.convert("RGBA")
-    character = resize_character(character, zoom)
-    character = crop_character(character, CHARACTER_BOX, offset)
-    return character
-
-
-def biggest_font_size(text: str, font_path, box_size, guess) -> ImageFont:
-    font = ImageFont.truetype(font_path, guess)
-    while (font.getsize(text)[0] > box_size[0]) or (font.getsize(text)[1] > box_size[1]):
-        guess -= 1
-        font = ImageFont.truetype(font_path, guess)
-
-    return font
-
-
 def draw_thumbnail_text(text: str, into: ImageType, center: tuple[int, int]) -> None:
-
     # multiple is added because rotating causes image to become jagged
     # sizing it down with the Image.ANTIALIAS flag smooths it out
     multiple = 3
@@ -97,9 +45,6 @@ def generate_thumbnail(data: TournamentData) -> ImageType:
     player_right = data["players"][1]
     canvas = Image.new("RGBA", SIZE)
 
-    # Output Path
-    output = os.path.join(OUTPUT_DIR, "test1.png")
-
     # TODO: This is allocating every time its run, make constant
     # Paths to Template Items
     background_path = os.path.join(THUMBNAIL_DIR, "Thumbnail-Background.png")
@@ -123,7 +68,7 @@ def generate_thumbnail(data: TournamentData) -> ImageType:
     # Player 1 Character
     canvas.alpha_composite(
         generate_character_image(
-            character_paths[0], player_left["offset"], player_left["zoom"]
+            character_paths[0], player_left["offset"], player_left["zoom"], CHARACTER_BOX
         ),
         POSITION[0],
     )
@@ -131,7 +76,7 @@ def generate_thumbnail(data: TournamentData) -> ImageType:
     # Player 2 Character
     canvas.alpha_composite(
         generate_character_image(
-            character_paths[1], player_right["offset"], player_right["zoom"]
+            character_paths[1], player_right["offset"], player_right["zoom"], CHARACTER_BOX
         ),
         POSITION[1],
     )
